Move pose trail metrics to new source file.
